start to refactor analyze for rancher to have multiple rancher modules
